requirements.txt
ffmpy
rich
pillow
numpy
setuptools
language_tool_python
pyyaml
questionary